+++ src/main/kotlin/lotto/process/LottoScore.kt
@@ -2,8 +2,8 @@ package lotto.process
 
 import lotto.model.LottoPrice
 import lotto.model.LottoPrize
+import lotto.model.LottoResult
 import lotto.model.LottoTicket
-import lotto.view.LottoResult
 
 class LottoScore {
     fun compareNumber(winningTicket: LottoTicket, lottoTickets: List<LottoTicket>): List<LottoResult> {
